chore: remove multi-generator (Compare/Recraft/adopt) feature (#56)
Removes compareGenerators/adoptGenerator actions, the Compare button,
the adopt UI and generator badges, the Recraft adapter, and
generateRecraftTransparent. Ideogram stays behind the adapter interface;
getGenerator falls back to it for historical recraft rows. DB columns
design_image.generator and design.active_generator_id untouched.
dedupeById moved to design-view.ts (compare.ts deleted).
