+++ src/app/design/actions.ts
@@ -17,7 +17,7 @@ import {
 import { eq, sql } from "drizzle-orm";
 import { chatAboutDesign, constructFluxPrompt, assessReadiness } from "@/lib/ai";
 import { uploadDesignImage, deleteDesignImageObject } from "@/lib/r2";
-import { getGenerator, GENERATORS, DEFAULT_GENERATOR_ID } from "@/lib/generators/registry";
+import { getGenerator } from "@/lib/generators/registry";
 import {
   insertDesignImage,
   reserveGenerationNumbers,
@@ -35,7 +35,7 @@ import {
 import { prefetchProductMockups } from "@/app/preview/actions";
 import { DEFAULT_BLANK_ID } from "@/lib/blanks";
 import { imageReferencedByOrders } from "@/lib/design-publish";
-import { compareSummary, dedupeById } from "@/lib/compare";
+import { dedupeById } from "@/lib/design-view";
 import type { ChatMessage } from "@/lib/db/schema";
 
 async function getOrCreateDesign(designId: string, userId: string) {
@@ -511,202 +511,4 @@ export async function approveDesign(designId: string) {
   // picks render instantly. Printful mockups are free so this is pure UX.
   // Best-effort: failures log and are swallowed by prefetchProductMockups.
   after(() => prefetchProductMockups(designId, DEFAULT_BLANK_ID));
-}
-
-/**
- * Run the same Claude-built prompt through every registered generator and
- * insert one design_image per result, tagged with its generator. Does NOT
- * change the design's active generator — that happens on adoptGenerator.
- * Returns the new images (id + url + generator) newest-last.
- */
-export async function compareGenerators(designId: string, userMessage?: string) {
-  const hdrs = await headers();
-  const session = await auth.api.getSession({ headers: hdrs });
-  if (!session) throw new Error("Unauthorized");
-
-  const found = await getOrCreateDesign(designId, session.user.id);
-  const ip = clientIp(hdrs);
-
-  // Abuse guard (#26 A3): Compare runs two paid model calls, so it's the
-  // costliest dead-click — count it against the daily caps before any work.
-  const quota = await consumeGenerationQuota({
-    userId: session.user.id,
-    isAnonymous: isAnonymousUser(session.user),
-    ip,
-  });
-  if (!quota.allowed) {
-    return { message: generationLimitMessage(quota.reason), images: [], readyToGenerate: true };
-  }
-
-  // Refund the consumed unit if the whole compare fails (see generateDesign).
-  try {
-    return await runCompare({ designId, userMessage });
-  } catch (err) {
-    await refundGenerationQuota({ userId: session.user.id, ip }).catch((e) =>
-      console.error("refundGenerationQuota failed:", e)
-    );
-    throw err;
-  }
-}
-
-async function runCompare({
-  designId,
-  userMessage,
-}: {
-  designId: string;
-  userMessage?: string;
-}) {
-  const messages = await getDesignMessages(designId);
-  const images = await getDesignImagesForAIContext(designId);
-  const messagesForPrompt: ChatMessage[] = userMessage
-    ? [...messages, { id: "pending", designId, role: "user", content: userMessage, imageId: null, createdAt: new Date() }]
-    : messages;
-
-  // Same fast thin-check as generateDesign — Compare runs two models, so a
-  // dead-click here is even costlier; bail in ~1s before any heavy work.
-  const readiness = await assessReadiness(messagesForPrompt, images, userMessage);
-  if (!readiness.ready) {
-    await persistClarification(designId, userMessage, readiness.question);
-    return { message: readiness.question, images: [], readyToGenerate: false, options: readiness.options };
-  }
-
-  let aiResponse;
-  try {
-    aiResponse = await constructFluxPrompt(messagesForPrompt, images, userMessage);
-  } catch (err) {
-    console.error("compareGenerators constructFluxPrompt failed:", err);
-    throw new Error("Failed to construct prompt");
-  }
-
-  if (isClarificationOnly(aiResponse.fluxPrompt)) {
-    await persistClarification(designId, userMessage, aiResponse.message);
-    return { message: aiResponse.message, images: [], readyToGenerate: false };
-  }
-
-  const anchorUrl =
-    aiResponse.referenceImage != null
-      ? images.find((img) => img.number === aiResponse.referenceImage)?.url
-      : undefined;
-
-  const gens = Object.values(GENERATORS);
-  // Reserve one number per adapter in a single atomic increment, so parallel
-  // uploads (here and across concurrent Compares/Generates) never collide on
-  // the R2 key designs/{id}/{n}.png. Failed adapters leave their number unused.
-  const reserved = await reserveGenerationNumbers(designId, gens.length);
-  // All compared images share the pre-compare latest source as their parent —
-  // anchoring on a fixed pick beats a "latest by createdAt" re-read that the
-  // parallel inserts would race.
-  const parentImageId = images[images.length - 1]?.id ?? null;
-
-  const results = await Promise.all(
-    gens.map(async (g, i) => {
-      try {
-        const url = await g.generate(g.adaptPrompt(aiResponse.fluxPrompt), {
-          aspect: "1:1",
-          referenceImageUrl: anchorUrl,
-          negativePrompt: aiResponse.negativePrompt,
-        });
-        const response = await fetch(url);
-        const buffer = Buffer.from(await response.arrayBuffer());
-        const r2Url = await uploadDesignImage(designId, reserved[i], buffer);
-        // Pre-generate the id and defer the DB insert to the batch below, so all
-        // rows commit atomically with the chat message + design update.
-        return {
-          imageId: crypto.randomUUID(),
-          imageUrl: r2Url,
-          generator: g.id,
-          cost: g.costPerImage,
-        };
-      } catch (err) {
-        console.error(`compareGenerators ${g.id} failed:`, err);
-        return null;
-      }
-    })
-  );
-
-  const ok = results.filter((r): r is NonNullable<typeof r> => r !== null);
-  if (ok.length === 0) throw new Error("All generators failed");
-
-  // Summarize honestly: name the styles that didn't come back rather than
-  // silently undercounting (#19 — "Compared 1 generators"). results is
-  // index-aligned with the generators we ran.
-  const succeeded = gens.filter((_, i) => results[i] !== null).map((g) => g.label);
-  const failed = gens.filter((_, i) => results[i] === null).map((g) => g.label);
-  const summary = compareSummary(succeeded, failed);
-
-  // Commit every image row + the summary message + the cost increment together.
-  // generationCount is already reserved above (atomic), so it's not written here;
-  // generationCost is an atomic increment over the successful adapters' cost.
-  const totalCost = ok.reduce((sum, r) => sum + r.cost, 0);
-  // db.batch is atomic, so element order is only for readability (no FK edges
-  // between these rows). Lead with the design update — a fixed first element
-  // keeps this a non-empty tuple for the batch type.
-  await db.batch([
-    db
-      .update(designTable)
-      .set({
-        generationCost: sql`${designTable.generationCost} + ${totalCost}`,
-        updatedAt: new Date(),
-      })
-      .where(eq(designTable.id, designId)),
-    ...ok.map((r) =>
-      db.insert(designImageTable).values({
-        id: r.imageId,
-        designId,
-        parentImageId,
-        aspectRatio: "1:1",
-        productId: null,
-        placementId: null,
-        imageUrl: r.imageUrl,
-        prompt: aiResponse.fluxPrompt,
-        generationCost: r.cost,
-        generator: r.generator,
-        isApproved: false,
-      })
-    ),
-    ...(userMessage
-      ? [
-          db.insert(chatMessageTable).values({
-            designId,
-            role: "user" as const,
-            content: userMessage,
-          }),
-        ]
-      : []),
-    db.insert(chatMessageTable).values({
-      designId,
-      role: "assistant" as const,
-      content: summary,
-    }),
-  ]);
-
-  return { message: summary, images: ok, readyToGenerate: true };
-}
-
-/**
- * Adopt a compared image: set the design's active generator to that
- * image's generator and make it the primary image. Owner-auth.
- */
-export async function adoptGenerator(designId: string, imageId: string) {
-  const session = await auth.api.getSession({ headers: await headers() });
-  if (!session) throw new Error("Unauthorized");
-
-  const found = await db.query.design.findFirst({ where: eq(designTable.id, designId) });
-  if (!found || found.userId !== session.user.id) throw new Error("Unauthorized");
-
-  const image = await db.query.designImage.findFirst({
-    where: eq(designImageTable.id, imageId),
-  });
-  if (!image || image.designId !== designId) throw new Error("Image not found");
-
-  await db
-    .update(designTable)
-    .set({
-      activeGeneratorId: image.generator ?? DEFAULT_GENERATOR_ID,
-      primaryImageId: imageId,
-      updatedAt: new Date(),
-    })
-    .where(eq(designTable.id, designId));
-
-  return { activeGeneratorId: image.generator ?? DEFAULT_GENERATOR_ID };
 }
