+++ src/app/design/page.tsx
@@ -5,8 +5,6 @@ import { useRouter, useSearchParams } from "next/navigation";
 import {
   sendChatMessage,
   generateDesign,
-  compareGenerators,
-  adoptGenerator,
   selectImage,
   deleteDesignImage,
   getDesign,
@@ -50,7 +48,6 @@ function DesignPageInner() {
   const [lightboxIndex, setLightboxIndex] = useState<number | null>(null);
   const [publishImageId, setPublishImageId] = useState<string | null>(null);
   const [drawerOpen, setDrawerOpen] = useState(false);
-  const [activeGenerator, setActiveGenerator] = useState("ideogram");
   // Soft nudge: Generate/Compare dim until Claude judges the idea concrete
   // (subject + style). A design with existing renders starts ready.
   const [readyToGenerate, setReadyToGenerate] = useState(false);
@@ -68,7 +65,6 @@ function DesignPageInner() {
         url: s.imageUrl,
         prompt: "",
         publishedAt: s.publishedAt,
-        generator: s.generator,
       }))
     );
     setProductGroups(productGroups);
@@ -107,7 +103,6 @@ function DesignPageInner() {
       if (!design) return;
       setMessages(chat);
       setSelectedImage(design.displayImageUrl);
-      setActiveGenerator(design.activeGeneratorId ?? "ideogram");
     });
     refreshGallery();
   }, [id, refreshGallery]);
@@ -186,43 +181,6 @@ function DesignPageInner() {
     } finally {
       setGenerating(false);
     }
-  }
-
-  async function handleCompare(userMessage?: string) {
-    setGenerating(true);
-    setOptions([]);
-    if (userMessage) {
-      setMessages((prev) => [...prev, makeOptimisticMessage("user", userMessage)]);
-    }
-    try {
-      await ensureGuestSession();
-      const result = await compareGenerators(designId.current, userMessage);
-      const { message, images: compared, readyToGenerate: ready } = result;
-      setMessages((prev) => [
-        ...prev,
-        makeOptimisticMessage("assistant", message),
-      ]);
-      setReadyToGenerate(ready);
-      setOptions("options" in result ? (result.options ?? []) : []);
-      // Empty when Claude asked a clarifying question instead of comparing.
-      if (compared.length) {
-        await refreshGallery();
-        if (window.matchMedia("(max-width: 767px)").matches) setDrawerOpen(true);
-      }
-    } catch {
-      setMessages((prev) => [
-        ...prev,
-        makeOptimisticMessage("assistant", "Comparison failed. Try again?"),
-      ]);
-    } finally {
-      setGenerating(false);
-    }
-  }
-
-  async function handleAdopt(imageId: string, imageUrl: string) {
-    const { activeGeneratorId } = await adoptGenerator(designId.current, imageId);
-    setActiveGenerator(activeGeneratorId);
-    setSelectedImage(imageUrl);
   }
 
   async function handleDeleteImage(imageId: string) {
@@ -338,8 +296,6 @@ function DesignPageInner() {
           generating={generating}
           onSend={handleSend}
           onGenerate={handleGenerate}
-          onCompare={handleCompare}
-          activeGenerator={activeGenerator}
           readyToGenerate={readyToGenerate}
           options={options}
           onUploadImage={handleUploadImage}
@@ -397,7 +353,6 @@ function DesignPageInner() {
           onDelete={handleDeleteImage}
           onMakeProducts={handleMakeProductsForImage}
           onPublish={handlePublishImage}
-          onAdopt={handleAdopt}
         />
       )}
