+++ src/lib/__tests__/ai.test.ts
@@ -574,7 +574,7 @@ describe("constructFluxPrompt", () => {
     });
 
     const images: DesignImage[] = [
-      { id: "img-1", number: 1, url: "https://example.com/1.png", prompt: "sunset", publishedAt: null, generator: null },
+      { id: "img-1", number: 1, url: "https://example.com/1.png", prompt: "sunset", publishedAt: null },
     ];
 
     const { constructFluxPrompt } = await import("../ai");
