+++ src/lib/design-images.ts
@@ -40,7 +40,6 @@ export type DesignImage = {
   url: string;
   prompt: string;
   publishedAt: Date | null;
-  generator: string | null;
 };
 
 /**
@@ -93,7 +92,6 @@ export async function getDesignImagesForAIContext(
     url: s.imageUrl,
     prompt: s.prompt ?? "",
     publishedAt: s.publishedAt,
-    generator: s.generator,
   }));
 }
 
@@ -110,7 +108,6 @@ export async function insertDesignImage(params: {
   generationCost: number;
   productId?: string | null;
   placementId?: string | null;
-  generator?: string | null;
   /** Explicit anchor. For placement renders (#25) this is the source image the
    * render was generated from, so a later lookup can match the exact pick.
    * Omitted for chat-driven generations → defaults to the latest thread image. */
@@ -138,7 +135,6 @@ export async function insertDesignImage(params: {
     imageUrl: params.imageUrl,
     prompt: params.prompt ?? null,
     generationCost: params.generationCost,
-    generator: params.generator ?? null,
     isApproved: false,
   });
   return id;
@@ -295,7 +291,6 @@ export type SourceImage = {
   prompt: string | null;
   createdAt: Date;
   publishedAt: Date | null;
-  generator: string | null;
 };
 
 /**
@@ -314,7 +309,6 @@ export async function getDesignSourceImages(
       prompt: designImageTable.prompt,
       createdAt: designImageTable.createdAt,
       publishedAt: designImageTable.publishedAt,
-      generator: designImageTable.generator,
     })
     .from(designImageTable)
     .where(
@@ -332,7 +326,6 @@ export async function getDesignSourceImages(
     prompt: r.prompt,
     createdAt: r.createdAt,
     publishedAt: r.publishedAt,
-    generator: r.generator,
   }));
 }
