+++ src/lib/generators/registry.ts
@@ -1,12 +1,10 @@
 import type { GeneratorId, ImageGenerator } from "./types";
 import { ideogramGenerator } from "./ideogram-generator";
-import { recraftGenerator } from "./recraft-generator";
 
 export const DEFAULT_GENERATOR_ID: GeneratorId = "ideogram";
 
 export const GENERATORS: Record<GeneratorId, ImageGenerator> = {
   ideogram: ideogramGenerator,
-  recraft: recraftGenerator,
 };
 
 /** Resolve an adapter by id, falling back to the default for null or
